requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
anyio==4.2.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.0
async-lru==2.0.4
async-timeout==4.0.2
attrs==22.2.0
Babel==2.14.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.1.0
cachetools==5.3.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.1
comm==0.1.4
contourpy==1.0.7
cycler==0.11.0
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
docker-pycreds==0.4.0
exceptiongroup==1.1.3
executing==2.0.0
fastjsonschema==2.19.0
filelock==3.10.4
Flask==2.2.3
fonttools==4.39.2
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.3.0
geomloss==0.2.6
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.16.3
google-auth-oauthlib==0.4.6
greenlet==3.0.3
grpcio==1.51.3
idna==3.4
importlib-metadata==6.1.0
importlib-resources==5.12.0
ipykernel==6.25.2
ipython==8.16.1
ipywidgets==8.1.1
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
julia==0.6.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==8.4.0
jupyter_core==5.4.0
jupyter_server==2.12.1
jupyter_server_terminals==0.5.0
jupyterlab==4.0.9
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
kaleido==0.2.1
kiwisolver==1.4.4
lightning-utilities==0.8.0
lit==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.8.2
matplotlib-inline==0.1.6
mistune==3.0.2
mpmath==1.3.0
multidict==6.0.4
munch==2.5.0
nbclient==0.9.0
nbconvert==7.13.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.0
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
opencv-python==4.7.0.72
overrides==7.4.0
packaging==23.0
pandas==2.0.0
pandocfilters==1.5.0
parso==0.8.3
pastel==0.2.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.11.0
playwright==1.43.0
plotly==5.21.0
poethepoet==0.10.0
prometheus-client==0.19.0
prompt-toolkit==3.0.39
protobuf==4.22.1
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyee==11.1.0
Pygments==2.16.1
pyparsing==3.0.9
pyppeteer==2.0.0
pysr==0.12.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytorch-lightning==2.0.0
pytz==2023.3
PyYAML==6.0
pyzmq==25.1.1
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.32.0
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.15.2
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
Send2Trash==1.8.2
sentry-sdk==2.7.1
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.3
style==1.1.0
sympy==1.11.1
tenacity==8.2.2
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
terminado==0.18.0
threadpoolctl==3.1.0
tinycss2==1.2.1
tomli==2.0.1
tomlkit==0.12.1
torch==2.0.0
torchcde==0.2.5
torchdiffeq==0.2.3
torchdyn==1.0.6
torchmetrics==0.11.4
torchsde==0.2.6
torchvision==0.15.1
tornado==6.3.3
tqdm==4.65.0
traitlets==5.11.2
trampoline==0.1.2
triton==2.0.0
types-python-dateutil==2.8.19.14
typing_extensions==4.5.0
tzdata==2023.3
update==0.0.1
uri-template==1.3.0
urllib3==1.26.15
wandb==0.17.3
wcwidth==0.2.8
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
websockets==10.4
Werkzeug==2.2.3
widgetsnbextension==4.0.9
yarl==1.8.2
zipp==3.15.0
